refactor(website): sanity structure display refactor
